one more attemp to fix error with session

+++ controllers/admin.js
@@ -1,6 +1,7 @@
 const Product = require('../models/product');
 const { validationResult } = require('express-validator');
 const fileHelper = require('../util/file');
+
 exports.getAddProduct = (req, res, next) => {
 	res.render('admin/edit-product', {
 		pageTitle: 'Add Product',
@@ -14,24 +15,11 @@ exports.getAddProduct = (req, res, next) => {
 
 exports.postAddProduct = (req, res, next) => {
 	const title = req.body.title;
+
 	const imageUrl = 'undefined';
 	const price = req.body.price;
 	const description = req.body.description;
-	// if (!image) {
-	// 	return res.status(422).render('admin/edit-product', {
-	// 		pageTitle: 'Add Product',
-	// 		path: '/admin/add-product',
-	// 		editing: false,
-	// 		hasError: true,
-	// 		product: {
-	// 			title: title,
-	// 			price: price,
-	// 			description: description,
-	// 		},
-	// 		errorMessage: 'Attached file is not an image. ',
-	// 		validationErrors: [],
-	// 	});
-	// }
+
 	const errors = validationResult(req);
 	if (!errors.isEmpty()) {
 		return res.status(422).render('admin/edit-product', {
@@ -103,6 +91,7 @@ exports.postEditProduct = (req, res, next) => {
 	const image = req.file;
 	const updatedDesc = req.body.description;
 	const errors = validationResult(req);
+
 	if (!errors.isEmpty()) {
 		return res.status(422).render('admin/edit-product', {
 			pageTitle: 'Edit Product',
@@ -119,6 +108,7 @@ exports.postEditProduct = (req, res, next) => {
 			validationErrors: errors.array(),
 		});
 	}
+
 	Product.findById(prodId)
 		.then(product => {
 			if (product.userId.toString() !== req.user._id.toString()) {
@@ -127,16 +117,20 @@ exports.postEditProduct = (req, res, next) => {
 			product.title = updatedTitle;
 			product.price = updatedPrice;
 			product.description = updatedDesc;
+
 			if (image) {
-				fileHelper.deleteFile(product.imageUrl);
-				product.imageUrl = image.path;
+				if (product.imageUrl && product.imageUrl !== 'undefined') {
+					fileHelper.deleteFile(product.imageUrl);
+				}
+
+				product.imageUrl = 'undefined';
 			}
+
 			return product.save().then(result => {
 				console.log('UPDATED PRODUCT!');
 				res.redirect('/admin/products');
 			});
 		})
-
 		.catch(err => {
 			const error = new Error(err);
 			error.httpStatusCode = 500;
@@ -147,7 +141,6 @@ exports.postEditProduct = (req, res, next) => {
 exports.getProducts = (req, res, next) => {
 	Product.find({ userId: req.user._id })
 		.then(products => {
-			console.log(products);
 			res.render('admin/products', {
 				prods: products,
 				pageTitle: 'Admin Products',
@@ -168,7 +161,10 @@ exports.deleteProduct = (req, res, next) => {
 			if (!product) {
 				return next(new Error('Product not found'));
 			}
-			fileHelper.deleteFile(product.imageUrl);
+
+			// if (product.imageUrl && product.imageUrl !== 'undefined') {
+			// 	fileHelper.deleteFile(product.imageUrl);
+			// }
 			return Product.deleteOne({ _id: prodId, userId: req.user._id });
 		})
 		.then(() => {
