+++ util/file.js
@@ -1,9 +1,14 @@
 const fs = require('fs');
+
 const deleteFile = filePath => {
+	if (!filePath || filePath === 'undefined') {
+		return;
+	}
 	fs.unlink(filePath, err => {
 		if (err) {
-			throw err;
+			console.log(err);
 		}
 	});
 };
+
 exports.deleteFile = deleteFile;
